Add an example illustrating the OPTIONS issue
https://github.com/douglasnaphas/aws-sam-cli/issues/1

To see the problem with OPTIONS described in this Issue:

  1. Make one directory where the branch for the PR addressing the issue
is checked out, from the awslabs/aws-sam-cli repo (the PR dir).
  1. Make another directory where the develop branch of the
awslabs/aws-sam-cli repo is checked out (the develop dir).
  1. In each of the above two dirs, activate a virtualenv for Python
3.7.4 and run `make init` to make a `samdev` executable available in
each virtualenv.
  1. In the PR dir, run `samdev local start-api --profile $PROFILE
--template $PLAY-SAM/dev-template.yml --port 5000 --host
localhost`
  1. In the develop dir, run `samdev local start-api --profile $PROFILE
--template $PLAY-SAM/dev-template.yml --port 5001 --host
localhost`
  1. In each dir, run `curl -X OPTIONS -i http://localhost:$PORT` where
`$PORT` is 5000 or 5001, depending on the dir.
  1. You should get a 200 with `"Output":"hello from options"}` as the
response data from the Express server in the PR dir, 200 with no data in
the develop dir. This illustrates the issue of OPTIONS not reaching the
handler.